backend/requirements.txt
requests>=2.28.0
argparse>=1.4.0
anthropic>=0.21.0
pandas>=1.5.3
python-dotenv>=1.0.0
PyPDF2>=3.0.0
scikit-learn>=1.0.0
numpy>=1.20.0